# Switch to canvas blobs for faster rendering