Use canvas.toBlob in place of canvas.toDataURL as an asynchronous way
to get image data from the canvas. Whammy.js is modified to support
this.

For reference, a single 30 sec, 60 fps replay benchmark on my computer rendered in 187 sec + 2 sec for compiling with existing canvas.toDataURL method compared to 46 sec + 3 sec for compiling with canvas.toBlob. The compiling time could be moved to a web worker in the future, if desired.

Whammy blob code taken from [the first jsfiddle](https://jsfiddle.net/121qq2ae/2/) in antimatter15/whammy#49.