@SomeBall-1 2016-12-31
I'll do some of the simple stuff. You're right about me not getting deep into Whammy code so I'll almost definitely need to leave that stuff up to you. I'll comment on change you listed.